Let attach FixedGuns everywhere; Move ProjectileData on a separate file
