+++ src/DynamicBody.h
@@ -77,17 +77,10 @@ public:
 		return tmp;
 	}
 
-	enum Feature {
-		PROPULSION = 0,
-		FIXED_GUNS = 1,
-		MAX_FEATURE = 2,
-	};
-
-	bool Have(Feature f) const { return m_features[f]; };
+	void AddFeature(Feature f) override;
 	void SetDecelerating(bool decel) { m_decelerating = decel; }
 	const Propulsion *GetPropulsion() const;
 	Propulsion *GetPropulsion();
-	void AddFeature(Feature f);
 
 protected:
 	virtual void SaveToJson(Json &jsonObj, Space *space) override;
@@ -118,8 +111,6 @@ private:
 	// for time accel reduction fudge
 	vector3d m_lastForce;
 	vector3d m_lastTorque;
-
-	bool m_features[MAX_FEATURE];
 
 	RefCountedPtr<Propulsion> m_propulsion;
 };
