+++ src/Projectile.cpp
@@ -3,6 +3,8 @@
 
 #include "Projectile.h"
 
+#include "ProjectileData.h"
+
 #include "CargoBody.h"
 #include "Frame.h"
 #include "Game.h"
